requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dnspython==2.5.0
email-validator==2.1.0.post1
Flask==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
PyMySQL==1.1.0
SQLAlchemy==2.0.27
typing_extensions>=4.13.2
Werkzeug==3.0.1
WTForms==3.1.2
greenlet==3.0.3

flask-security-too==5.3.3
flask-admin==1.6.1
flask-marshmallow==1.2.0
# bcrypt==4.1.2
marshmallow-sqlalchemy==1.0.0
bcrypt==3.2.2

# -- Nuevos requierements seguridad (Login y 2FA) --
setuptools<70.0.0
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-Mailman==1.1.1
passlib==1.7.4
cryptography==46.0.6
cffi==2.0.0
pycparser==3.0
qrcode==8.2
pillow==12.1.1
phonenumbers==9.0.26
